Splits the Service from Excel Grids and the templates fro Data and Summary in different classes.
