+++ src/main/java/com/werfen/report/model/ReportHeaderConfiguration.java
@@ -3,10 +3,10 @@ package com.werfen.report.model;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.ArrayList;
 import java.util.List;
-
-import static java.util.Objects.nonNull;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Builder
 @Getter
@@ -19,16 +19,6 @@ public class ReportHeaderConfiguration {
     private final GridReportField field4;
 
     public List<GridReportField> getFields() {
-        List<GridReportField> fields = new ArrayList<>();
-        if (nonNull(this.getField1()))
-            fields.add(this.field1);
-        if (nonNull(this.getField2()))
-            fields.add(this.field2);
-        if (nonNull(this.getField3()))
-            fields.add(this.field3);
-        if (nonNull(this.getField4()))
-            fields.add(this.field4);
-
-        return fields;
+        return Stream.of(field1, field2, field3, field4).filter(Objects::nonNull).collect(Collectors.toList());
     }
 }
